backend/requirements.txt
flask
google-generativeai
faiss-cpu
PyPDF2
numpy
